chore: fixed api to add a new connection
