+++ src/controllers/connection.controller/index.js
@@ -1,3 +1,4 @@
 module.exports.getConnection = require('./getConnection');
 module.exports.acceptConnection = require('./acceptConnection');
 module.exports.rejectConnection = require('./rejectConnection');
+module.exports.addConnection = require('./addConnection');
